stdint fix

+++ libc/v1.0/include/stdint.h
@@ -5,7 +5,7 @@
 typedef char int8_t;
 typedef short int16_t;
 typedef int int32_t;
-typedef long int64_t;
+typedef long long int64_t;
 typedef char int_least8_t;
 typedef short int_least16_t;
 typedef int int_least32_t;
@@ -13,25 +13,25 @@ typedef long int_least64_t;
 typedef char int_fast8_t;
 typedef short int_fast16_t;
 typedef int int_fast32_t;
-typedef long int_fast64_t;
-typedef long intmax_t;
-typedef long intptr_t;
+typedef long long int_fast64_t;
+typedef long long intmax_t;
+typedef long long intptr_t;
 
 // unsigned integer types
 typedef unsigned char uint8_t;
 typedef unsigned short uint16_t;
 typedef unsigned int uint32_t;
-typedef unsigned long uint64_t;
+typedef unsigned long long uint64_t;
 typedef unsigned char uint_least8_t;
 typedef unsigned short uint_least16_t;
 typedef unsigned int uint_least32_t;
-typedef unsigned long uint_least64_t;
+typedef unsigned long long uint_least64_t;
 typedef unsigned char uint_fast8_t;
 typedef unsigned short uint_fast16_t;
 typedef unsigned int uint_fast32_t;
-typedef unsigned long uint_fast64_t;
-typedef unsigned long uintmax_t;
-typedef unsigned long uintptr_t;
+typedef unsigned long long uint_fast64_t;
+typedef unsigned long long uintmax_t;
+typedef unsigned long long uintptr_t;
 
 //  int max limits
 #define INTMAX_MIN -9223372036854775807
